requirements.txt
pytelegrambotapi
python-dotenv
requests
loguru
